Convert formik to functional component + react-query
